requirements.txt
Flask==2.1.2
Pillow==8.4.0